+++ tests-drafts/EndpointParameters/ConnectorSlugTest.php
