+++ tests-drafts/EndpointParameters/StringParameterTest.php
