+++ tests-drafts/RegistrarTest.php
