+++ tests-drafts/Registrars/ConnectorRegistrarTest.php
